+++ scifit/tests/test_solvers_report.py
@@ -5,7 +5,6 @@ from unittest import TestCase
 from scifit.solvers import linear, scientific, specials
 from scifit.toolbox import report
 
-
 path = pathlib.Path(".cache/media/reports")
 path.mkdir(parents=True, exist_ok=True)
 
@@ -13,7 +12,6 @@ print_report = bool(int(os.getenv("TESTS_PRINT_REPORT", 1)))
 
 
 class TestBasicReportProcessor(TestCase):
-
     file = "report_linear"
     context = {
         "title": "Fit Report",
@@ -32,7 +30,11 @@ class TestBasicReportProcessor(TestCase):
 
     def test_pandoc_converter(self):
         if print_report:
-            self.processor.convert("# Hello world\n\n##Foo\n\n###Bar", file="dummy", path=".cache/media/reports/")
+            self.processor.convert(
+                "# Hello world\n\n##Foo\n\n###Bar",
+                file="dummy",
+                path=".cache/media/reports/",
+            )
 
     def test_serialize_figure(self):
         axe = self.solver.plot_fit()
@@ -44,19 +46,36 @@ class TestBasicReportProcessor(TestCase):
 
     def test_report_pdf(self):
         if print_report:
-            self.processor.report(self.solver, context=self.context, path=".cache/media/reports", file=self.file, mode="pdf")
+            self.processor.report(
+                self.solver,
+                context=self.context,
+                path=".cache/media/reports",
+                file=self.file,
+                mode="pdf",
+            )
 
     def test_report_docx(self):
         if print_report:
-            self.processor.report(self.solver, context=self.context, path=".cache/media/reports", file=self.file, mode="docx")
+            self.processor.report(
+                self.solver,
+                context=self.context,
+                path=".cache/media/reports",
+                file=self.file,
+                mode="docx",
+            )
 
     def test_report_html(self):
         if print_report:
-            self.processor.report(self.solver, context=self.context, path=".cache/media/reports", file=self.file, mode="html")
+            self.processor.report(
+                self.solver,
+                context=self.context,
+                path=".cache/media/reports",
+                file=self.file,
+                mode="html",
+            )
 
 
 class TestBadReportConstant(TestBasicReportProcessor):
-
     file = "report_constant"
     factory = linear.ConstantFitSolver